DDA-103: Refactor useDeleteDioryAndLinks to its own file
